refactor: menuMessage 파일로  분리
