+++ src/js/common/utils.js
