+++ src/js/menu/menuMessage.js
@@ -1,0 +1,24 @@
+const menuMessage = {
+  espresso: {
+    title: "☕ 에스프레소 메뉴 관리",
+    placeholder: "에스프레소 메뉴 이름",
+  },
+  frappuccino: {
+    title: "🥤 프라푸치노 메뉴 관리",
+    placeholder: "프라푸치노 메뉴 이름",
+  },
+  blended: {
+    title: "🍹 블렌디드 메뉴 관리",
+    placeholder: "블렌디드 메뉴 이름",
+  },
+  teavana: {
+    title: "🫖 티바나 메뉴 관리",
+    placeholder: "티바나 메뉴 이름",
+  },
+  desert: {
+    title: "🍰 디저트 메뉴 관리",
+    placeholder: "디저트 메뉴 이름",
+  },
+};
+
+export default menuMessage;
